Move mDNS workflow out from wifi_setup.cpp
