+++ src/wifi_setup.h
@@ -10,7 +10,6 @@
 
 
 void initWiFi();
-void mdnsReset();
 
 void disconnectWiFi();
 void reconnectWiFi();
